nrt: reserve: clean up code layout
Split the trivial Passthrough cache implementation in its own file,
and keep only interfaces only cleanly separated.
